Refactor: PlaceList JSX엘리먼트 컴포넌트화
